create MessagePopup and MessagePopupStyle, update UserProfile display UI of friends or trangers
