+++ app/App.js
@@ -21,6 +21,7 @@ import UserProfile from "./screens/profile/userProfile/UserProfile"; // Đảm b
 import AddFriend from "./screens/profile/userProfile/AddFriend"; // Đảm bảo đường dẫn đúng
 import AcceptFriend from "./screens/profile/userProfile/AcceptFriend"; // Đảm bảo đường dẫn đúng
 import Options from "./screens/optional/OptionalScreen";
+
 const Stack = createNativeStackNavigator();
 
 function RootLayout() {
@@ -55,11 +56,7 @@ function RootLayout() {
         <Stack.Screen name="UserProfile" component={UserProfile} />
         <Stack.Screen name="AddFriend" component={AddFriend} />
         <Stack.Screen name="AcceptFriend" component={AcceptFriend} />
-        <Stack.Screen
-          name="Options"
-          component={Options}
-          options={{ animation: "slide_from_right" }} // Thêm hiệu ứng xếp chồng lên
-        />
+        <Stack.Screen name="Options" component={Options} options={{ animation: "slide_from_right" }}/>
       </Stack.Navigator>
     </PaperProvider>
   );
