+++ package.json
@@ -24,22 +24,26 @@
     "@react-navigation/stack": "^7.1.2",
     "@rneui/base": "^4.0.0-rc.7",
     "@rneui/themed": "^4.0.0-rc.8",
-    "expo": "^52.0.37",
+    "expo": "~52.0.38",
+    "expo-barcode-scanner": "^13.0.1",
     "expo-blur": "~14.0.3",
-    "expo-constants": "17.0.7",
+    "expo-camera": "^16.0.18",
+    "expo-constants": "~17.0.8",
     "expo-font": "13.0.4",
     "expo-haptics": "~14.0.1",
     "expo-image-picker": "~16.0.6",
     "expo-linear-gradient": "~14.0.2",
     "expo-linking": "7.0.5",
-    "expo-router": "~4.0.17",
+    "expo-modules-core": "^2.2.3",
+    "expo-permissions": "^14.4.0",
+    "expo-router": "~4.0.19",
     "expo-splash-screen": "0.29.22",
     "expo-status-bar": "~2.0.1",
     "expo-symbols": "0.2.2",
     "expo-system-ui": "4.0.8",
     "expo-web-browser": "~14.0.2",
     "fuse.js": "^7.1.0",
-    "jest-expo": "52.0.5",
+    "install": "^0.13.0",
     "link": "^2.1.1",
     "moment": "^2.30.1",
     "react": "18.3.1",
@@ -57,7 +61,7 @@
     "react-native-paper": "^5.13.1",
     "react-native-phone-number-input": "^2.1.0",
     "react-native-reanimated": "~3.16.1",
-    "react-native-safe-area-context": "^4.14.1",
+    "react-native-safe-area-context": "^4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-tab-view": "^4.0.5",
     "react-native-vector-icons": "^10.2.0",
@@ -70,7 +74,7 @@
     "@types/react": "~18.3.12",
     "@types/react-test-renderer": "^18.3.0",
     "jest": "^29.2.1",
-    "jest-expo": "~52.0.5",
+    "jest-expo": "~52.0.6",
     "react-test-renderer": "18.3.1",
     "typescript": "^5.3.3"
   },
